requirements.txt
streamlit>=1.31.0
google-generativeai>=0.3.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0 